refactor: Moves EventSerializer to EventStore as DefaultEventSerializer.
