+++ src/Tutor.Web/Startup.cs
@@ -20,12 +20,13 @@ using Tutor.Core.LearnerModel.Feedback;
 using Tutor.Core.LearnerModel.Notes;
 using Tutor.Infrastructure;
 using Tutor.Infrastructure.Database.EventStore;
+using Tutor.Infrastructure.Database.EventStore.DefaultEventSerializer;
 using Tutor.Infrastructure.Database.Repositories;
 using Tutor.Infrastructure.Database.Repositories.Domain;
 using Tutor.Infrastructure.Database.Repositories.Learners;
+using Tutor.Infrastructure.EventConfiguration;
 using Tutor.Infrastructure.Security;
 using Tutor.Infrastructure.Security.Authentication;
-using Tutor.Infrastructure.Serialization;
 using Tutor.Web.Controllers.Domain.DTOs.AssessmentItems.ArrangeTasks;
 using Tutor.Web.Controllers.Domain.DTOs.AssessmentItems.Challenges;
 using Tutor.Web.Controllers.Domain.DTOs.AssessmentItems.MultiResponseQuestions;
@@ -67,7 +68,12 @@ namespace Tutor.Web
                 registry.RegisterType<TextDto>();
                 registry.RegisterType<VideoDto>();
 
-                serializerOptions.SetupEvents();
+                registry.RegisterConvention(new AllowedTypesDiscriminatorConvention<string>(
+                    serializerOptions, EventSerializationConfiguration.EventRelatedTypes, "typeDiscriminator"));
+                foreach (var type in EventSerializationConfiguration.EventRelatedTypes.Keys)
+                {
+                    registry.RegisterType(type);
+                }
             });
 
             services.AddCors(options =>
@@ -93,11 +99,11 @@ namespace Tutor.Web
 
             services.AddScoped<IFeedbackService, FeedbackService>();
             services.AddScoped<IFeedbackRepository, FeedbackDatabaseRepository>();
-            
+
             services.AddScoped<INoteRepository, NoteRepository>();
             services.AddScoped<INoteService, NoteService>();
 
-            services.AddScoped<IEventSerializer, EventSerializer>();
+            services.AddSingleton<IEventSerializer>(new DefaultEventSerializer(EventSerializationConfiguration.EventRelatedTypes));
 
             SetupAuth(services);
